Add files via upload
In this project, I stretched myself  a bit and researched on how to visualize biological data in R. Using scatter plot, histogram, box plots, I plotted two vectors as well as improved it aesthetically by adding color, point in character (pch), title, and label. I also used mtext and paste function to insert the correlation coefficient of the vectors in the graph.